WL#2412 (Online Backup : Kernel updates for object metadata changes):
This is a preliminary patch introducing changes outlined in the worklog. The 
main changes include:

- Introduction of Backup_restore_ctx class and moving big parts of 
preparation/clean-up code there. Refactoring execute_backup_command() to use it.

- Refactoring in Image_info class and friends. Catalogue items allocated using 
mem_root. Use of Map datatype for storing position -> object mappings.

- Moving all reporting functionality into the Logger class (including reporting
to the online backup progress tables).

One thing which is missing is reorganization of the source code - move class
definitions to separate files and consistent use of #include directive. This will
be done in a separate patch.

Also missing is correctly reporting error messages in all places 
where errors are detected. This involves going through all the places marked 
with "TODO: report errors" comment and also places where temporary 
ER_BACKUP_UNKNOWN_ERROR is reported and define and report appropriate error 
messages there.

+++ sql/backup/api_types.h
@@ -187,10 +187,10 @@ class Table_list
     virtual ~Table_list() {}
 
     /// Return reference to given list element. Elements are counted from 0.
-    virtual Table_ref operator[](uint pos) const =0;
+    virtual Table_ref operator[](ulong pos) const =0;
 
     /// Return number of elements in the list.
-    virtual uint  count() const =0;
+    virtual ulong count() const =0;
 };
